requirements.txt
asgiref==3.7.2
Django==4.2.4
Pillow==10.0.0
python-environ==0.4.54
sqlparse==0.4.4
typing_extensions==4.7.1
